Keep *Interpreter methods in interp.go; error.go holds only the error type
error.go now mirrors types/error.go: it declares RuntimeError, FrameInfo,
escape, and the sentinel errors, plus RuntimeError's own Error/Unwrap. The
*Interpreter methods that build and dispatch them (runtimeError, guard,
cause, handle, handler, land, discard, wrap, uncaught, message,
framesInfo) move back to interp.go, since a type's methods don't relocate
away from where the rest of that type lives.
